Add support for tags which are not SemVer compliant
Since the version info is fetched over git tags, the tag names
had to be SemVer compliant strings before this commit.
This fact did break the version fetching process when
tags were set for anything else. This commit fixes this
issue. All kind of tags are allowed now, as long as they
are not in the form of: v0abc.4def.2ghi