clean up hal cmake
move the conditional inclusion of sources bassed on the 'USE_*' flags
into the add_driver_option function for better transparency. The headers
are still installed unconditionally. This leaves the possibility of
linker errors caused by including headers whose source files have not
been compiled. This is not a new issue, but this also doesn't fix that.

It also adds proper handling for non-cmake libraries in add_driver_option
assuming that all these libraries support pkg-config.