fix examples not being installed correctly
Some of the examples relied on (misnamed) _FOUND variables to only
include certain examples if a library was found (presumably to check if
certain HW support was enabled) to conditionally add examples. This was
broken (or in some cases maybe hasn't been working right for some time)
by the naming changes for targets & dependencies (e.g. canopenlib).
This commit refactors the exampls' CMakeLists.txt to be more consistent,
readable and correct.