add TimeSource to centrally handle timekeeping
Not only does this provide a signle point of truth for all things time,
but it also allows for much better testing using the ManualTime source
that does not rely on passing real time. This also decouples the Executor
from potential time/sync sources like ROS and EtherCAT.

Note that this is still a work in progress and ROS & EtherCAT time
sources are not yet implemented.